+++ src/bda/plone/shop/interfaces.py
@@ -2,6 +2,7 @@ from zope.interface import Interface
 from zope import schema
 from zope.i18nmessageid import MessageFactory
 from zope.interface import alsoProvides
+from plone.supermodel import model
 from bda.plone.cart.interfaces import ICartItem
 from plone.directives import form
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
@@ -35,7 +36,7 @@ class IShopSettingsProvider(Interface):
     """
 
 
-class IShopSettings(form.Schema):
+class IShopSettings(model.Schema):
     """Shop controlpanel schema.
     """
 
@@ -64,11 +65,11 @@ class IShopSettings(form.Schema):
             'bda.plone.shop.vocabularies.CurrencyDisplayOptionsVocabulary')
 
 
-class IShopCartSettings(form.Schema):
+class IShopCartSettings(model.Schema):
     """Shop controlpanel schema for cart settings.
     """
 
-    form.fieldset(
+    model.fieldset(
         'cart',
         label=_(u'Cart Settings'),
         fields=[
@@ -108,11 +109,11 @@ class IShopCartSettings(form.Schema):
 alsoProvides(IShopCartSettings, IShopSettingsProvider)
 
 
-class IShopArticleSettings(form.Schema):
+class IShopArticleSettings(model.Schema):
     """Shop controlpanel schema for article settings.
     """
 
-    form.fieldset(
+    model.fieldset(
         'article',
         label=_(u'Article Settings'),
         fields=[
@@ -168,11 +169,11 @@ class IShopArticleSettings(form.Schema):
 alsoProvides(IShopArticleSettings, IShopSettingsProvider)
 
 
-class IShopShippingSettings(form.Schema):
+class IShopShippingSettings(model.Schema):
     """Shop controlpanel schema for article settings.
     """
 
-    form.fieldset(
+    model.fieldset(
         'shipping',
         label=_(u'Shipping Settings'),
         fields=[
@@ -198,11 +199,11 @@ class IShopShippingSettings(form.Schema):
 alsoProvides(IShopShippingSettings, IShopSettingsProvider)
 
 
-class IShopTaxSettings(form.Schema):
+class IShopTaxSettings(model.Schema):
     """Shop controlpanel schema for tax settings.
     """
 
-    form.fieldset(
+    model.fieldset(
         'tax',
         label=_(u'Tax Settings'),
         fields=[